Update on "example of splitting up an FX graph into smaller subgraphs with own submodules"
Differential Revision: [D23956147](https://our.internmc.facebook.com/intern/diff/D23956147)

[ghstack-poisoned]
